Massive refactor to get descriptions working.
Completely changes the way templates are fetched and stored.  Must be rolled out at the same time as the backend changes that go along with it.

Repositories are now keyed by `provider`/`repo_id` instead of `owner`/`name`.  This means that we have to go to GitHub to get metadata.

We may want to consider piping GitHub requests through a proxy we control so we can use our keys for un-authenticated users.  Unauthenticated users may hit their cap relatively quickly if the number of sponsored/popular repositories is high since each one requires a query for GitHub metadata.  We could also cache if rate limiting is a problem.

Deleted all old versions of the app, they won't work with the new backend anyway and it was getting confusing trying to figure out what was what.